requirements.txt
# Claude Auto SEO + Digital Marketing — Python Dependencies

# Core HTTP & scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0

# Image generation (banner creator)
Pillow>=10.0.0

# NLP & text analysis
nltk>=3.8.1
textstat>=0.7.3

# Data & machine learning
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0

# Google integrations (Analytics, Search Console, GMB)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Social media
tweepy>=4.14.0

# DataForSEO
dataforseo-client>=1.0.0

# PDF generation
reportlab>=4.0.0
weasyprint>=60.0

# WordPress
python-wordpress-xmlrpc>=2.3

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.6.0
tabulate>=0.9.0
jinja2>=3.1.2
schedule>=1.2.0
